requirements.txt
numpy-1.23.5
PyGithub-1.57 
certifi-2022.12.7 
cffi-1.15.1 
charset-normalizer-3.0.1 
deprecated-1.2.13 
idna-3.4 
pycparser-2.21 
pyjwt-2.6.0 
pynacl-1.5.0 
requests-2.28.2 
urllib3-1.26.14 
wrapt-1.14.1
pyyaml-6.0
